# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * YouTube Content Studio API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type IdeaStatus = typeof IdeaStatus[keyof typeof IdeaStatus];


export const IdeaStatus = {
  NEW: 'NEW',
  RESEARCHING: 'RESEARCHING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
  SCRIPTING: 'SCRIPTING',
  READY: 'READY',
} as const;

export type IdeaSource = typeof IdeaSource[keyof typeof IdeaSource];


export const IdeaSource = {
  MANUAL: 'MANUAL',
  STRATEGIST: 'STRATEGIST',
} as const;

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  IDEA: 'IDEA',
  RESEARCH: 'RESEARCH',
  SCRIPT: 'SCRIPT',
  SCENES: 'SCENES',
  MEDIA: 'MEDIA',
  VOICE: 'VOICE',
  EDIT: 'EDIT',
  QUALITY_CHECK: 'QUALITY_CHECK',
  READY: 'READY',
  UPLOADED: 'UPLOADED',
  PUBLISHED: 'PUBLISHED',
  FAILED: 'FAILED',
} as const;

export type ScriptStatus = typeof ScriptStatus[keyof typeof ScriptStatus];


export const ScriptStatus = {
  DRAFT: 'DRAFT',
  APPROVED: 'APPROVED',
} as const;

export interface Idea {
  id: string;
  title: string;
  topic: string;
  hook: string;
  description: string;
  targetAudience: string;
  format: string;
  estimatedDuration: number;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  viralScore: number | null;
  /**
     * @minimum 0
     * @maximum 100
     * @nullable
     */
  originalityScore: number | null;
  /** @nullable */
  rationale: string | null;
  weaknesses: string[];
  improvements: string[];
  source: IdeaSource;
  status: IdeaStatus;
  createdAt: string;
  updatedAt: string;
}

export interface IdeaInput {
  /**
     * @minLength 1
     * @maxLength 180
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  topic: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  hook: string;
  /** @maxLength 3000 */
  description: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  targetAudience: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  format: string;
  /**
     * @minimum 10
     * @maximum 180
     */
  estimatedDuration: number;
}

export interface IdeaUpdate {
  /**
     * @minLength 1
     * @maxLength 180
     */
  title?: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  topic?: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  hook?: string;
  /** @maxLength 3000 */
  description?: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  targetAudience?: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  format?: string;
  /**
     * @minimum 10
     * @maximum 180
     */
  estimatedDuration?: number;
  status?: IdeaStatus;
}

export interface GenerateIdeasInput {
  /** @maxLength 200 */
  topic?: string;
  /** @maxLength 240 */
  targetAudience?: string;
  /** @maxLength 100 */
  format?: string;
  /**
     * @minimum 1
     * @maximum 8
     */
  count?: number;
}

export interface AnalyzeIdeaInput {
  ideaId?: string;
  /** @maxLength 180 */
  title?: string;
  /** @maxLength 200 */
  topic?: string;
  /** @maxLength 500 */
  hook?: string;
  /** @maxLength 3000 */
  description?: string;
  /** @maxLength 240 */
  targetAudience?: string;
  /** @maxLength 100 */
  format?: string;
  /**
     * @minimum 10
     * @maximum 180
     */
  estimatedDuration?: number;
}

export interface StrategistAnalysis {
  /**
     * @minimum 0
     * @maximum 100
     */
  viralScore: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  originalityScore: number;
  explanation: string;
  hookOptions: string[];
  weaknesses: string[];
  improvements: string[];
}

export interface ProjectInput {
  ideaId: string;
  /**
     * @minLength 1
     * @maxLength 180
     */
  name?: string;
}

export interface ProjectUpdate {
  /**
     * @minLength 1
     * @maxLength 180
     */
  name: string;
}

export interface ProjectStatusUpdate {
  status: ProjectStatus;
}

export interface ContentProject {
  id: string;
  ideaId: string;
  name: string;
  status: ProjectStatus;
  ideaTitle: string;
  scriptCount: number;
  createdAt: string;
  updatedAt: string;
}

export interface SceneInput {
  /** @minimum 1 */
  sequence: number;
  narration: string;
  visualInstructions: string;
  onScreenText: string;
  soundSuggestion: string;
}

export type ScriptScene = SceneInput & {
  id: string;
  scriptId: string;
};

export interface ShortsScript {
  id: string;
  projectId: string;
  ideaId: string;
  hook: string;
  narration: string;
  soundMusic: string;
  cta: string;
  status: ScriptStatus;
  scenes: ScriptScene[];
  createdAt: string;
  updatedAt: string;
}

export interface ProjectDetail {
  project: ContentProject;
  idea: Idea;
  scripts: ShortsScript[];
}

export interface GenerateScriptInput {
  ideaId: string;
  projectId?: string;
  /**
     * @minimum 15
     * @maximum 180
     */
  targetDurationSeconds?: number;
  /** @maxLength 120 */
  tone?: string;
}

export interface ScriptUpdate {
  /**
     * @minLength 1
     * @maxLength 800
     */
  hook?: string;
  /**
     * @minLength 1
     * @maxLength 12000
     */
  narration?: string;
  /** @maxLength 1000 */
  soundMusic?: string;
  /** @maxLength 800 */
  cta?: string;
  /** @maxItems 24 */
  scenes?: SceneInput[];
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface PerformanceItem {
  videoTitle: string;
  views: number;
  likes: number;
  comments: number;
  publishedAt: string;
  url: string;
}

export interface DashboardSummary {
  videosCreated: number;
  videosPublished: number;
  ideasGenerated: number;
  scriptsCreated: number;
  projectsInProduction: number;
  ideasByStatus: StatusCount[];
  productionByStatus: StatusCount[];
  performanceAvailable: boolean;
  recentPerformance: PerformanceItem[];
  recentProjects: ContentProject[];
  recentIdeas: Idea[];
}

/**
 * Invalid input
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Resource not found
 */
export type NotFoundResponse = ErrorResponse;

/**
 * Request conflicts with current state
 */
export type ConflictResponse = ErrorResponse;

/**
 * The configured AI provider is unavailable
 */
export type AIUnavailableResponse = ErrorResponse;

export type ListIdeasParams = {
status?: IdeaStatus;
/**
 * @maxLength 160
 */
q?: string;
};

export type ListProjectsParams = {
status?: ProjectStatus;
};

export type ListScriptsParams = {
projectId?: string;
};