# lib/api-zod/src/generated/types/ideaInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * YouTube Content Studio API
 * OpenAPI spec version: 0.1.0
 */

export interface IdeaInput {
  /**
     * @minLength 1
     * @maxLength 180
     */
  title: string;
  /**
     * @minLength 1
     * @maxLength 200
     */
  topic: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  hook: string;
  /** @maxLength 3000 */
  description: string;
  /**
     * @minLength 1
     * @maxLength 240
     */
  targetAudience: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  format: string;
  /**
     * @minimum 10
     * @maximum 180
     */
  estimatedDuration: number;
}