# lib/api-zod/src/generated/types/projectDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * YouTube Content Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ContentProject } from './contentProject';
import type { Idea } from './idea';
import type { ShortsScript } from './shortsScript';

export interface ProjectDetail {
  project: ContentProject;
  idea: Idea;
  scripts: ShortsScript[];
}