# lib/api-zod/src/generated/types/scriptScene.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * YouTube Content Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { SceneInput } from './sceneInput';

export type ScriptScene = SceneInput & {
  id: string;
  scriptId: string;
};