# lib/api-zod/src/generated/types/shortsScript.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * YouTube Content Studio API
 * OpenAPI spec version: 0.1.0
 */
import type { ScriptScene } from './scriptScene';
import type { ScriptStatus } from './scriptStatus';

export interface ShortsScript {
  id: string;
  projectId: string;
  ideaId: string;
  hook: string;
  narration: string;
  soundMusic: string;
  cta: string;
  status: ScriptStatus;
  scenes: ScriptScene[];
  createdAt: Date;
  updatedAt: Date;
}